add zshenv(export EDITOR)

+++ setup.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 # .zshrc
-curl https://raw.githubusercontent.com/sup39/zsh-setup/master/zshrc -o $HOME/.zshrc
+[ -f $HOME/.zshrc ] || curl https://raw.githubusercontent.com/sup39/zsh-setup/master/zshrc -o $HOME/.zshrc
+# .zshenv
+[ -f $HOME/.zshenv ] || curl https://raw.githubusercontent.com/sup39/zsh-setup/master/zprofile -o $HOME/.zshenv
+
 # .vim/sup39
 [ -d $HOME/.vim/sup39 ] || git clone https://github.com/sup39/vimrc $HOME/.vim/sup39
 # .vimrc
